requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
openai==1.30.1
scikit-learn==1.5.0
nltk==3.8.1
numpy==1.26.4
pandas==2.2.2
httpx==0.27.0
python-dotenv==1.0.1
pytest==8.2.2
pytest-asyncio==0.23.7
